+++ addons/behaviortree/BT_Node.cs
@@ -8,7 +8,7 @@ public abstract partial class BT_Node : Node
 {
     public enum BT_Status { Success, Failure, Running }
     
-    public virtual BT_Status OnBehave(float delta, Node owner)
+    public virtual BT_Status OnBehave(double delta, Node owner)
     {
         return BT_Status.Failure;
     }
