+++ addons/behaviortree/BT_Selector.cs
@@ -5,7 +5,7 @@ using Godot;
 [GlobalClass, Tool]
 public partial class BT_Selector : BT_Composite
 {
-    public override BT_Status OnBehave(float delta, Node owner)
+    public override BT_Status OnBehave(double delta, Node owner)
     {
         foreach (BT_Node node in Children)
         {
