+++ addons/behaviortree/BT_Sequence.cs
@@ -5,7 +5,7 @@ namespace AutoCrawler.addons.behaviortree;
 [GlobalClass, Tool]
 public partial class BT_Sequence: BT_Composite
 {
-    public override BT_Status OnBehave(float delta, Node owner)
+    public override BT_Status OnBehave(double delta, Node owner)
     {
         foreach (BT_Node node in Children)
         {
